Yet Another DFAO Optimization
I give up. This effect is just too unstable to use in conjunction with Lumen. It makes the game look way too dark.

It is viable to use standalone, and GI Low has that setup. It doesn't look very good, and compared to just turning on Lumen, the performance hit is not worth it.

I'm still leaving this here since I spent way too much time trying to make this work. But I'm never touching this setting again.